+++ lib/empy_widget.dart
@@ -3,18 +3,19 @@ import 'package:flutter/material.dart';
 /// Widget destinado aos casos nos quais a tela deveria contem informações úteis
 /// mas que, por algum motivo, não estão disponíveis.
 class EsigEmptyWidget extends StatelessWidget {
-  const EsigEmptyWidget(this.mensagem,{
+  const EsigEmptyWidget(
+    this.mensagem, {
     Key? key,
     this.ilustracaoAsset,
     this.tamanhoIlustracao,
     this.titulo,
     this.tituloFontSize,
     this.tituloPadding = const EdgeInsets.all(8.0),
-    
     this.mensagemFontSize,
     this.mensagemPadding = const EdgeInsets.all(8.0),
     this.altura,
     this.largura,
+    this.corTexto,
   }) : super(key: key);
 
   /// Caminho do asset declarado no [pubspec.yaml]. A imagem está configurada
@@ -41,6 +42,8 @@ class EsigEmptyWidget extends StatelessWidget {
 
   /// Padding em torno da mensagem.
   final EdgeInsets mensagemPadding;
+
+  final Color? corTexto;
 
   /// Altura total do widget. Parâmetro serve para evitar usar o código abaixo
   /// toda vez que implementar este Widget.
@@ -94,7 +97,7 @@ class EsigEmptyWidget extends StatelessWidget {
           titulo!,
           textAlign: TextAlign.center,
           style: TextStyle(
-            color: Colors.black,
+            color: corTexto,
             fontSize: tituloFontSize,
             fontWeight: FontWeight.bold,
           ),
@@ -106,7 +109,7 @@ class EsigEmptyWidget extends StatelessWidget {
         child: Text(
           mensagem,
           textAlign: TextAlign.center,
-          style: TextStyle(fontSize: mensagemFontSize),
+          style: TextStyle(color: corTexto, fontSize: mensagemFontSize),
         ),
       );
 }